Minor documentation and comment updates. Break YarnShuffleAlternateNameConfigSuite into a separate class. Fix new test in YarnShuffleServiceSuite when run with other tests in the suite.
